refactor: degree 패키지 분리
