+++ src/test/java/woowacrew/article/domain/ArticleConverterTest.java
@@ -3,7 +3,7 @@ package woowacrew.article.domain;
 import org.junit.jupiter.api.Test;
 import woowacrew.article.dto.ArticleResponseDto;
 import woowacrew.article.utils.ArticleConverter;
-import woowacrew.user.domain.Degree;
+import woowacrew.degree.domain.Degree;
 import woowacrew.user.domain.User;
 
 import static org.assertj.core.api.Assertions.assertThat;
