+++ src/test/java/woowacrew/security/provider/SocialLoginAuthenticationProviderTest.java
@@ -10,8 +10,8 @@ import woowacrew.oauth.OauthService;
 import woowacrew.oauth.github.GithubOauthService;
 import woowacrew.security.token.SocialPostAuthorizationToken;
 import woowacrew.security.token.SocialPreAuthorizationToken;
-import woowacrew.user.domain.Degree;
-import woowacrew.user.domain.DegreeRepository;
+import woowacrew.degree.domain.Degree;
+import woowacrew.degree.domain.DegreeRepository;
 import woowacrew.user.domain.User;
 import woowacrew.user.domain.UserRepository;
 import woowacrew.user.dto.UserContext;
