+++ src/test/java/woowacrew/user/domain/UserTest.java
@@ -1,5 +1,6 @@
 package woowacrew.user.domain;
 
+import woowacrew.degree.domain.Degree;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import woowacrew.user.domain.exception.ForbiddenUserException;
